combine the two extensions, delete the old unused one
