--- contrib-nbextension/.gitignore
@@ -1,15 +1,0 @@
-*.egg-info/
-.ipynb_checkpoints/
-dist/
-build/
-*.py[cod]
-node_modules/
-
-# Compiled javascript
-beaker-nbextension/static/
-
-# OS X
-.DS_Store
-/.classpath
-/.project
-/.settings/
