--- contrib-nbextension/beaker.yaml
@@ -1,7 +1,0 @@
-Type: IPython Notebook Extension
-Name: Beaker
-Description: Part of BeakerX, see http://BeakerNotebook.com
-Link: README.md
-Icon: icon.png
-Main: main.js
-Compatibility: 4.x
