--- contrib-nbextension/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "beaker",
-  "version": "0.0.0",
-  "private": true,
-  "dependencies": {
-    "bower": "1.3.8"
-  }
-}
